app/requirements.txt
psycopg2-binary==2.9.3
requests==2.32.3
datetime==5.5
beautifulsoup4==4.13.3
fake-useragent==1.4.0

